minor update, some fixes in entity objects, ui fixes

+++ src/main/java/com/rwr/service/IRwrManagementService.java
@@ -4,8 +4,8 @@ import com.rwr.entity.ims.ImsType;
 import com.rwr.entity.seeker.Seeker;
 import com.rwr.entity.skills.SkillType;
 import com.rwr.utils.IPageWrapper;
-import com.rwr.utils.PageWrapperImpl;
 import com.rwr.utils.Pageable;
+
 import java.util.List;
 import java.util.Set;
 
@@ -29,5 +29,5 @@ public interface IRwrManagementService {
 
     void delete(Integer id);
 
-    Integer getSeekerRowCount();
+    Long getSeekerRowCount();
 }
